fix: Dynamically parameterized tests name matching
Dynamically parameterized tests (the ones not using [InlineData] in
xUnit for example) have generated values in intermediate_test_results
table, which are not present in statically generated tree nodes. This
mismatch causes mapping to fail when generating neo test results and
ultimately reports test run as a failure even when it succeeds.

Resolves #40