feat: shorten redundant part of displayname (#6)
This commit checks the files once read from vstest if the DisplayName
and FullyQualifiedName start with the same part. If they do, we trim the
DisplayName up to (and including) the last period.

This achieves a compact view of the test without a lot of namespace
redundancy in the neotest tree.